requirements.txt
flask==2.0.2
flask-cors==3.0.10
gunicorn==20.1.0
redis==4.0.2
python-socketio==5.5.2